move sections into tabs
